vecbook/requirements.txt
fastmcp>=0.1.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
toml>=0.10.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
requests>=2.31.0
